[dcl.fct.def.general] fix extraneous space in __func__ definition ('"function-name "' -> '"function-name"') and elsewhere.
This fixes all instances of \term{} and \grammarterm{} used in code
blocks to not insert spaces afterwards. I fixed the one instance where
a space was required and not already explicitly present, and checked
all the rest to ensure they are as they were before or better. Many
(mostly harmless but ugly) extraneous spaces in code blocks containing
italicized terms are now gone.