N3649 Generic (Polymorphic) Lambda Expressions (Revision 3)
Because of conflicts with N3638, I had to change the [dcl.spec.auto]
changes quite a bit, and I recommend CWG create an issue to review the
merged state.

In the example provided (f1, f2, g, h, glambda), it is not clear to me
what "ID" refers to in comments like "error: ID is not convertible".