LWG2093 Throws clause of condition_variable::wait with predicate
I edited the first sentence of the new paragraph in
[thread.req.timing] for improved grammar and clarity.

I introduced the term "timeout-related exceptions" to make it clear
what is referred to by that phrase later on.

I dropped "any" from "any timeout-related exceptions" in some of the
Throws clauses modified to by the proposed wording to be consistent
with the other Throws clauses.