N3778 Sized Deallocation
Editorial fixes to the wording:

[expr.delete]/9
Included the (not highlighted as being added) wording starting with "Otherwise" as it
seemed to be intended to be added to complete the paragraph.

[new.delete.array]/13
Changed the signatures to drop the nothrow_t parameter in this paragraph as this
section deals with throwing array deallocation functions, not non-throwing ones.

[diff.cpp11.basic]
Added a reference to [basic.stc.dynamic.deallocation] for the change description.
Reworded the Effect paragraph to match other change descriptions.
Fixed reference "new.expr" to "expr.new".