P0012R1 Make exception specifications be part of the type system, version 5
The change to [except.spec] 15.4p1 could not be applied, but the current
wording is correct. Editorially fixed formatting errors in wording,
reformatted text for consistency and readability, and fixed use of
macros.

The change in the formulation of function pointer decay to avoid
suggesting that only possibly-throwing functions decay has been applied
to two more places in [except.handle].