[unique.ptr.special] prefer use of common_type_t
The preferred style since C++14 is to use the _t alias
rather than the trait::type formulation for transformation
traits.  As a second fix, the types in the common_type
instantiation are, in turn, dependant types so require a
leading typename keyword.  I double-checked elsewhere and
we are consistent to use typename where requrired inside
other expressions, even though it is frequently omitted
when the named type is used within the surrounding
English text.