[tuple.helper] prefer to use tuple_element_t (#686)
Prefer to use the _t alias to the old trait::type formulation
when specifying the place-holder name for defining
tuple_element of a cv-qualified type.  This neatly sidesteps
the question of whether there is a missing typename in the
subsequent usage, or whether that would be implied in the
use of the place-holder.