deque should not reference vector
Deque introduces itself in terms of vector, to describe the idea of
random access iterators.  This is perhaps more confusing that simply
directing the reader to the clause for random access iterators.  For
C++17 we introduce the category of contiguous iterators for vector,
array, and string, and we do not want to confuse the reader into bad
assumptions.  Meanwhile, the notion of iterators is much more widely
understood than in 1998, so pointing directly to the appropriate
clause is more likely to help the reader than a potentially misleading analogy.