[meta.unary.prop] Consistent formatting for 'void' (#684)
* [meta.unary.prop] Consistent formatting for 'void'

Fixes #544.

Update all use of 'void' in type traits tables to use the codified
form \tcode{void}.  There was one use of the phrase "void types"
that I deliberately chose to not touch.

A second issue I stumbled over in this edit was inconsistent use
of cv-qualified vs. \cv-qualfied.  The latter seemed preferable,
so I applied that formatting consistently through this file for
all uses of cv-qualified that were not otherwise participating
in mark-up.

* Fix formatting of cv-qualifiers