[allocator.adaptor] consolidate memory utilities]
This patch simply moves the scoped allocator clause to
be adjacent to the other memory and allocator clauses
for a more consistent sub-organization of clause 20.

The patch looks pretty awful thanks to the diff
algorithm treating it as a lot of interspersed edits,
where in fact is way a single cut/paste for the bulk
of the test (plus a second one-liner for the clause 20
table of contents).