[string.view.comparison] Fix index for basic_string_view operators
The first issue is that operator<< does not group in the index with
other definitions for operator<<, which use an index key of
operator\shl.

The second issue is that operator<< was the only operator indexed
under the 'basic_string_view' key, where the precedent fro basic_string
and other types is that the free-function operators are still indexed
under their respective class entry.