[18-28] Fix index entry for constructors
The preferred way to list constructors in the index is under the
non-code font term 'constructor', but many classes retain old
text that lists the constructor under the class name as the same
class name, in a code font.  Worse, there are quite a few classes
that have fallen into a hybrid approach, with both listings, which
can be confusing. This change-set should catch every constructor
indexed in the old style and consistently transform it to the new.

One additional fix is that allocator_traits mis-indexed the static
member functions 'construct' and 'destroy' as the plain-text
constructor and destructor.  This patch correctly indexes them as
their correct name, in a code-font.