[18-27] Remove tcode from index entries
Several old index entries use '\tcode' to force a code
font on their sub-entry.  This is no longer necessary,
as the LaTeX scripts produce the correct font for an
'\idxcode' entry.  However, using '\tcode' inconsistently
will produce duplicate entries, and entries that do not
sort correctly.  This change set consistently applies
'\idxcode' in preference to '\tcode' to eliminate the
redundant and mis-sorted entries.