[meta.trans.other] Reformat Other Transformations table
This PR applies the change recommended by https://github.com/cplusplus/draft/issues/629.

So far, I like the change giving the descriptions more room to
breathe, but for come reason the first column has become a
little narrower, which is messing with the format a little.

Generally, I think this looks like a good fix, but may use a
little more polish to get the presentation just right.