[18,20] Consistent indexing of simple exception types
Appply a simple, consistent pattern to the index entries, including
th index of implementation defined behaviors, for the library types
that derive from std::exception.

In some cases this means adding entries, in others removing extra
nested implementation-defined entries that are redundant since the
addition of the index of implementatio defined behaviors.

A few cross-references were added or corrected as part of this
process.

Further work to make the presentation and specification of these
types probably requires LWG issues.  In particular, bad_any_cast
is woefully underspecified.