[18-20,27] Prefer to use 'override' where appropriate]
This patch makes consistent use of the 'override' contextual
keyword for every member function (other than destructors)
that overrides a virtual function in one of its base classes.

In general, this came down to one of three cases:
1) the 'what' function for an exception class
2) the allocation functions of a memory resource
3) the implementation methods of a stream buffer.

As far as I can tell, all other virtual function declarations
in the standard library are intended as customization points
for users, and are not overriden in the standard itself.