[strings,containers] Add iterators to index of implemenetation defined behavior (#754)
This patch slavishly copies the tag used for array::iterator, without
understanding if a better markup might be available.  One consequence
is that we get two entries for basic_string_view::const_iterator.
Another is that the index around type names appears as two sorted
subsequences, rather than one sorted sequence.

An additional patch will follow for the remaining implementation-defined
types that are not present in the index, once the preferred markup is
reviewed in this pull request.