[19-26] Consistent use of _v
The library clauses have started to make inconsistent use of the _v
variable templates for traits, in preferene to directly querying
trait::value.  This change set applies that change consistently over
all existing wording, where there is a choice of which form might
be used.  It does not make any changes where the _v variable is not
aviable, or is being defined in terms of the corresponding ::value.