[hardware.interference] prefer itemdecl to codeblock
The new hardware interference constants are introduced
above the definition of their semantics with a codeblock,
rather than an itemdecl, macro.  Prefer the latter for
consistently introducing names that are about to be defined,
as this will simply auditing the text in the future, e.g.,
when trying to reivew that every itemdecl has a corresponding
library index entry.