[language.support] review library indexing of clause 18
This change consistently applies to new indexlibrarymember
macro in places missed by the first pass - mostly for cases
where either the class or member-name has an underscore.

It consistently uses a % to eliminate potential whitespace
being introduced by an index entry macro.

It removes a duplicate entry for 'numeric_limits' causes
by the use of the \_ command.  This appears to be the
correct fix, on inspection of the index.  The alternative
would have been consistent use of \_ on every numeric_limits
entry, so trying the simpler fix first.