Audit index of implementation-defined behavior
Review all uses of the terms 'implementation-defined' and
'implementation defined' in the standard, and replace with
\impldef entries in the index of implementation defined
behavior where appropriate.

Clean up some older index entries that did not have a clear
reference, and appear to predate the index of implementation
defined behavior being added by Pete Becker for C++11.

Changes may appear more disruptive than in practice, due to
word-wrapping reflowing a paragraph or two in the doc source,
but not on the rendered pdf.