[depr, exception.syn] Consistently move deprecated declarations to Annex D]
This change moves the deprecated declarations in the <exception>
header to Annex D, in a manner consistent with the library pattern
of fully specifying deprecated components and names in the deprecation
annex.

Then apply consistent wording across the library parts of the annex
describing how non-deprecated headers introduce the deprecated extensions.

Finally, ensure that the deprecated extensions to standard headers are
indexed as part of that standard header.