[lex.phases] Index and xref raw string literals (#902)
Add a cross-reference on the reversion of universal characters in raw string
literals, as it is far from clear that [lex.pptoken] is the place to look
for this rule, which is not spelled out clearly in the phase1/phase2 rules.
Remove a confusing pair of parentheses as it was not clear if the intent
was to make the parenthetical a note, which we have a better way to render,
or normative.  Given the requirement that universal characters behave
consistently, this seems normative, rather than a note.

Index raw string literals, which appear to be entirely lacking from the
main index.