[meta] Add index entries for each type trait
Create an index entry for each row in the type traits tables,
indexing the corresping trait.  Where a trait is defined outside
the table, add a second index reference to the latter.  This
causes an annoying duplication, as the current software sees the
index entry inside the table as in some way NOT the same as the
entry outside the table.

Disambiguate the is_empty function from the filesystem library,
and the is_empty trait.  The issue for is_signed being a trait
and a member of numberic_limits resolves itself.