[diff.cpp14.special] Clarify example comment
The comment describing the reason for the new behavior was subject to
misinterpretation; a reader may have thought that the substition
failure was the cause of A(int) being called, which is not the case.
This patch reorders the clauses of the sentence so that they follow
a logical cause-and-effect order.