NB US-91 (C++17 CD): [algorithm][numeric] Specify all parallel algorithms
Also addresses US-157, US-183

Add a copy of the parallel algorithm signature below each corresponding
non-parallel signature in the specification for each algorithm in the
<algorithm> and <numeric> headers.  This is *mostly* an exercise in
copy/paste - however a small subset of algorithms have wording that
either refers to the signatures obliquely, requiring a minor wording
tweak, or uses the 'Effects: as if ...' formulation, which requires a
separate specification for the parallel form to perfectly forward the
execution policy.

This change corresponds to (known to be pending) national body comments
requesting more clarity in the specification of each algorithm that has
a parallel overload.

A further tweak disambiguates the indexing of the various 'move'
functions.