[atomics] Clean up indexing of atomics clause
Reverse class/member names in indexlibrarymember macros to be
consistent with the prefered style in other clauses.

Expand in the index several functions that are documented as a
pattern-match, such as fetch_add and fetch_sub.

Replace 'atomic type' for index references with either just 'atomic',
'atomic<integral>', or 'atomic<T*>' to follow existing conventions
for documenting templates, and to more clearly call out the larger
interfaces of the defined specializations.

Added further indexing for a few items that had missed index entries
in the first pass.