[17-30] Create index entrie for all namespaces
Index each library namespace, and the reserved posix namespace.
Note that many entries for literals namespaces are added after
the header syonpsis, as typically they are the last entries in
that synposis, and the index will find them much more easily
here.  Indexing inside the synopsis is avoided, as it leads to
duplicate index entries.