[macros] Make 'itemize' environments evenly spaced, decrease itemdecl indent.
Each item is now separated from other items and from ambient paragraphs by the same amount as paragraphs are separated from one another. This is also true for paragraphs within an item, and for nested lists.

This was achieved by modifying the 'itemize' and 'enumerate' environments to set the relevant lengths appropriately, and by resetting the depth counter to 0 for every list, regardless of actual nesting level, to overcome hardcoded scaling logic in the fundamental 'list' environment.

Lists are now spaced independently of whether they are preceded by a paragraph break (because partopsep is now zero).

Also changes the 'itemdecl' environment to not use lists for indentation, but 'adjustwidth' instead, and decreased the indentation to the same as for code listings.