[styles] Give lists uniform vertical spacing.
This change makes it so that the vertical space surrounding lists is independent of whether there is a paragraph break before the list (by setting partopsep to zero).

Moreover, all nested lists now use the same vertical spacing, which in turn is the same as the paragraph spacing.

There is a bug in memoir that may cause second-level lists (lists within a top-level list) to be rendered too tightly (because the itemsepii length is ignored). This can be fixed locally, though.