removes uses of term functor
Fixes issue #1403 by removing all mentions of the term "functor". While
there is a convention to use "functor" to mean "function object" it
seems to be fading away. Also, the word "functor" has very different
meaning in Category Theory and (as a consequence of that) in many
Functional Programming languages. As a result of that use of the word
"functor" in the Standard could be confusing to the reader.