Consistent comma after e.g. and i.e.
After grepping to determine the preferred idiom, it seems clear that
the standard prefers to follow both e.g. and i.e. with a comma, rather
than omit it.  This patch applies that rule to the few places that
were missing the comma, including in LaTeX comments only for the editors'
benefit.