P0298R3 A byte type definition
Removed some unnecessary commas in [dcl.init], and reordered
cross-references to clarify that they apply to both the unsigned narrow
character type case and the std::byte case.

Updated the introductory sentence in [cstddef.syn] to explicitly state
that std::byte and its operations are not part of the corresponding C
header. Reordered detailed descriptions of operations to match order of
declarations in synposis. Replaced "IntegerType" with "IntType" to fix
nasty line break problem.

[basic.life] Did not remove still-necessary "or".