P0623R0 Final C++17 Parallel Algorithms Fixes
Reordered one list to make more a more logical pattern.
Did not add the footnote, as the previous patch provided
a global note covering this in the front-matter for this
subclause.

This patch is applied on top of p0574r1, and is the sixth
and final patch in this set of parallel algorithm edits.