Consistently use 'operator""x' rather than 'operator "" x' throughout standard.
We must remove the second space for 'operator""if', because
'operator "" if' is not valid, and should have a single consistent
style. Evidence suggests that people either put spaces both before
and after the "" or in neither place, so the only uniform consistent
common style is this one.

Fixes wrong index entry for 'operator""if' and bad collation.