Replace "this International Standard" with "this document" in some places where it is a clear improvement.
The 7th Edition of the ISO Directives no longer require us to use
"this International Standard" when the document refers to itself. We
retain that phrasing when the reference is to the abstract notion of
the specification as opposed to the text embodying it.

This is a step towards addressing #1389.