[Editorial] Clarify UTF-8 character literal definition
The definition of UTF-8 character literals almost paralleled that of
char16_t character literals, but there were a few differences; this
commit removes those.

One of the differences in particular was that the specification of the
value of a UTF-8 character literal was missing the phrase "containing a
single c-char", which is necessary both as the antecedent of "its" later
in the sentence.