Reword [over.ics.rank]/3.2.3
The wording here is awkward. The intent is, I think, to prefer binding an rvalue to an rvalue reference than to an lvalue reference, but the phrase "to an rvalue" is attached to the rvalue reference part (to which it is redundant). Would be clearer to attach it to the lvalue reference part.