Use 'one of' formulation for binary-digit
The grammar for literals consistently prefers using the 'one of' a set
formulation for character sets, rather than listing each character as
an option.  This holds for other parts of this grammar, even when the
set of options is just two items, such as hexadecimal-prefix,
unsigned-suffix, or long-suffix.

As binary-digit is the odd one out, this simple update makes it
consistent with the rest of the grammar specifications in this
clause.