Improved typesetting of probability functions.
By using the new `\p` and `\P` macros, probability functions will be typeset
more consistently and cleanly.  The syntax of both macros is the same.
Examples:

    \p{z}
    \p{z | a, b}
    \p*{z | \frac{1}{2}}
    \p[\big]{\frac{A}{B}}

The pipe (|) will have the proper spacing automatically.
The `\p*` variant will autoscale the parentheses and vertical bar to fit the contents of the expressions.
You can specify a particular size using the `\p[<size>]` variant.
Possible sizes are `\big`, `\Big`, `\bigg`, and `\Bigg`.