[dcl.mptr] Add index entry for pointer to member clause
There is already an index entry for 'pointer to member', but
it does not point to 11.3.3 [dcl.mptr] that provides the rules
for the declarator.  This bit me just today when I could not
find the magic words to find just this clause, so adding an
extra reference under the existing index entry.