[span.overview] fix reverse_iterator for span
There is a common convention to omit 'std::' when not strictly needed,
other than for 'std::move' and 'std::forward'.  A third case is for
'reverse_iterator' and 'const_reverse_iterator' type aliases inside a
class definition, as the leading 'reverse_iterator' type alias hides
the 'std' version from the 'const_reverse_iterator' definition, and
subsequently fails to compile.  This patch restore the class definition
to a safely copy/pastable state.

An additional review indicated this convention has been applied correctly
for every othert potential occurrence in the library.