[diff.cpp03.language.support] Remove change described in p1, which is no longer a possible change.
It is not possible to write a global replacement operator new or
operator delete in C++20 with code that is also valid C++03, because

a) operator new must be declared 'throw(std::bad_alloc)' in C++03 and
   must be declared with no exception specification in C++11 onwards,
and
b) operator delete must be declared 'throw()' in C++03 and must be
   declared 'noexcept' in C++20 onwards.

Therefore there is no code that is valid in C++03 and C++20 and changes
meaning due to the change identified in this section.

Instead, expand [diff.cpp03.language.support]p2 to explain that the
affected programs are simply not valid in C++20.