[expr.prim.id.dtor] Fix comment in example.
In translation phase 3, the "0.T" in the line of code is parsed as a (single) preprocessing-token.

In phase 7, this preprocessing token is converted into a user-defined-floating-literal whose ud-suffix component is "T".

At no point is the "0." partial token a floating-point literal. At most, it is the fractional-constant component in a bigger user-defined-floating-literal token.

Fixes #2630.