[alg.random.sample] Remove redundant requirement
Per [rand.req.urng]/1, the result type of a `UniformRandomBitGenerator` must model `UnsignedIntegral`, i.e., it must be an integral type. [alg.random.sample]/1.5 requires `Distance` to be an integer type. Since all integer types are convertible to all other integer types, the result type of any `UniformRandomBitGenerator` is convertible to `Distance`. The library need not enforce the rules of the core language.