[temp.deduct] Fix Ambiguous "this" in paragraph 3
This is an editorial change. There is an ambiguous "this"
in paragraph 3 of [temp.deduct]. Upon first reading this paragraph
it is not obvious that "this substitution" refers to the
substitution process described in the subsequent paragraphs.
This change specifies that the substitution process refers
to the one described in the subsequent paragraphs of the section.

Note: The example in that paragraph was expanded to use a codeblock
because prevent and overfull hbox.