[intro.refs] ISO/IEC 2382 missing date
Undated references are mentioned to indicate the latest edition. However, the latest edition of ISO/IEC 2382 is from 2015 (see https://www.iso.org/standard/63598.html), and the only place in the standard which mentions ISO/IEC 2382 explicitly is [intro.defs] and it mentions it along with the edition from 1993.

So it seems like either the 1993 edition should be explicitly listed in [intro.refs] (this is the suggested fix in this pull request) or the document (or at least the explicit reference in [intro.defs]) should be updated to mention the 2015 edition (assuming there were no breaking changes since 1993, which I definitely did *not* verify).