[concepts.arithmetic] fix notes that use undefined terms
The terms "signed integral types" and "unsigned integral types" are not
defined in [basic.fundamental] The notes are trying to talk about
signed/unsigned *integer* types. char and bool are not signed or
unsigned *integer* types, but they certainly are *integral* types, and
so they model one of signed_integral or unsigned_integral.