[extern.types] strike footnote listing C stdlib types
This footnote seemed like it was trying to exhaustively list all types imported from the C stdlib, but it was incomplete. Most notably, it was missing the types from `<stdint.h>`. Removing rather than fixing since the list is of questionable value and would be likely to become stale again.