[dcl.fct] Fix reference to temp.fct.
This reference was originally spelled "17.6.5" in the wording paper P1141R2,
where it was relative to the Concepts TS, and the error slipped through.
During motion application, we spotted the error, but not the cause, and
picked temp.over.link as a "best-fit" replacement, when in fact the correct
reference is temp.fct.