[atomic] Minor corrections to atomic index
The index for the primary atomic template was split across
two entries, atomic and atomic<T>.  Reconciled into one.

A couple of operations for atomic smart pointers are
specificed in the primary template, but were missing the
cross-reference into that specification, which is supplied
in the same way for all the other atomic specializations.