[expr.ref] clarify access of non-static data members
According to the definitions, objects don't have members, and member subobjects don't have names. The standard is somewhat unclear on how the member subobjects of an object of class type relate to the non-static data members of the class. This change helps clarify that there is a member subobject associated with each non-static data member.