[support.dynamic] Improve cross-referencing for dynamic memory subclauses
Update several cross-references to subclause 17.6 with more precise
refernces to subclauses nested one level deeper.  For the synopsis
of the header, added a couple of banner comments with cross-references
too.

In addition, looked into ordering the global namespace declarations
in front of opening namespace std to improve clarity, but that fails
due to use of types declared in namespace std by this header!