[expr] Clarify performing / applying standard conversions.
We previously had a conflict between wording demanding that certain
conversions be applied, and the definitions of the conversions saying
they could not be applied. We now distinguish two cases:

 * When we say that a conversion is applied, we're talking about a
   change to the resulting model of the program -- we mean that the
   conversion actually happens (and it had better be applicable).
 * When we say that a conversion is performed, we're talking about
   performing a process, which applies the conversion if possible and
   otherwise has no effect.

In addition, wording saying to perform the three kinds of conversion
from glvalue to prvalue have been changed to say that a prvalue operand
is expected.