[intro.refs,time.format] Fix normative references.
ISO/IEC/IEEE 60559:2011 and UAX#29 are not normative
requirements of C++ and thus were moved to the bibliography.
For ISO 8601:2004, highlighted its normative impact on
time formatting.
Move the footnote about the Unicode trademark to the
new first mention of the term in [intro.memory].

Fixes ISO/CS 002 (C++20 DIS)