[everywhere] Stop talking about C++ International Standards.
As far as ISO is concerned, there is only one International Standard for
C++, and in any case, we don't mean the document here, we mean the
language in the abstract and don't care whether that's an ISO standard
or not. Refer to "revisions of C++" instead of revisions of particular
ISO documents.

Partially addresses ISO/CS 016 (C++20 DIS).