Clarify definition of "variable"
In the current definition, the special exception for non-static data members is only stated for references, which gives the misleading impression that non-reference non-static data member declarations are variable declarations. By re-wording the definition, we can make it clear that a non-static data member declaration is never a variable declaration.