[meta.member] Clarify is_corresponding_member semantics
The definition of 'common initial sequence' in [class.mem] only applies
to standard-layout struct types, which excludes unions. There is no
reason to define is_corresponding_member in terms of standard-layout
types (which includes unions and scalars and arrays of such types) when
the common initial sequence is only meaningful for standard-layout
structs.