variant - clarifying the meaning of Ti
We preface some of the sections with "In the descriptions that follow, let i be in the range [0, sizeof...(Types)), and Ti be the ith type in Types.", but we do not do this for all the sections that use Ti. Similarly, there is another place where we use TI, but do not elaborate on what it means.