[macros] In \ucode, align size of digits and letters (#5211)
* [macros] In \ucode, improve sizes of digits and letters.

We do this by typesetting letters in small-caps of the body font,
and digits using the font "smaller by 1 step" (e.g. 10pt -> 8pt).
This combination seems to look perfectly harmonious for
normal body text and notes.

This works around the absence of "small-caps sized lining numerals"
in Latin Modern. The fact that LMR has separate optical sizes for the
size range in question makes this combination work well.