[iterator.operations] clarify preconditions on std::distance
The prior wording implied that the two sides of the "or" were independent, so as long as `last` was reachable from `first`, `first` didn't need to be reachable from `last` even if passing a `RandomAccessIterator`. It seems clear that it is really trying to express two different possible preconditions depending on whether a `RandomAccessIterator` is passed or not. I used the same phrasing as in the "Effects:" clause to separate the cases.